screenlock: enable rtc wakeup source
Please apply to the crust feature branch. The rtc wakeup clock is useful cause it allows
users to script programs. If I put my phone into deep sleep, none of the scripts
I put in my cronjob will run.

For example, if I have an alarm script:

https://git.sr.ht/~anjan/sxmo-userscripts/tree/master/alarm.sh

It is supposed to run at 7:00 am but since I put my phone in deep sleep before going to sleep,
the script doesnt play sound to wake me up.

Enabling the rtc wakeup clock enables using programs like rtcwake as a wrapper around
your cronjobs to wake up the phone at a time and excecute the command you put in cron.

For info on how to use rtcwake, see:
https://wiki.postmarketos.org/wiki/Nokia_N900_(nokia-n900)#Setting_the_hardware_alarm_clock

Not sure if we should patch busybox cron to use rtcwake automatically or whether we should
trust the user to use rtcwake properly and document it. However, I am certain that the rtc
wakeup source should be enabled for this usecase.

Signed-off-by: Miles Alan <m@milesalan.com>