added deeper context menus for vim/vis, nano, tuir, ncmpcpp, w3m
This patch looks at whatever application is running inside the st terminal.
and presents a more fine-tuned context menu based on it. It does so by
parsing WM_NAME.

Supported for the moment: vim/vis, nano, tuir, ncmpcpp, w3m

This probably won't work if you're in a terminal multiplexer like
screen/tmux, unless it's configured to propagate the name of running
application.

This patch is to be applied **after** my patch for feh/sxiv from earlier today (not included in this patch)

Signed-off-by: Miles Alan <m@milesalan.com>