Enhanced application/scripts menu
I added a whole bunch of possibly useful tools to the
application menu (will only show if they are actually installed) and sorted
the list alphabetically. Also included a new hook so the user can completely
override the app menu. The scripts and user-scripts menus are now merged into
one.