modemmonitor: use dbus-monitor
Instead of looping over mmcli commands to check for incoming
calls and texts every 3s, use dbus-monitor to tell us when
something is coming in.  This should save some battery life
as well as reduce the delay in new call/msg notification.

I was running on the latest released scripts, so have not
strictly tested this with git master.  Note that I dropped
the call to checkmissedcalls.  Perhaps we should still run
that at firststart?

Signed-off-by: Serge Hallyn <serge@hallyn.com>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>