sxmo_xinit: add support for overriding vars with device-specific profiles
This reads base/compatible from the devicetree to get the device name,
and sources a script in /bin with a filename in the format
'sxmo_deviceprofile_NAME.sh'. The purpose of this is to allow
setting/overriding variables in other sxmo scripts with device-specific
values. For example, some audio devices on one phone may have different
names on another.

Signed-off-by: Anjandev Momi <anjan@momi.ca>