sxmo_deviceprofile_librem5*: add initial profile for librem5
Currently there are 3 revisions of the L5, and they are all sharing the
same profile here since this only sets audio vars and they're all the
same. In the future the symlinks can be broken if profiles need to
deviate for some reason in order to set different values for vars, etc.

Signed-off-by: Anjandev Momi <anjan@momi.ca>