sxmo_brightness: add support for overriding backlight device
This uses the BACKLIGHT env var to override the device used for
controlling the backlight. My intention was to keep the default behavior
if it's unset, use the backlight/backlight path if that exists then fall
back to backlight/edp-backlight.

Finally, this changes the script to return immediately if no backlight
device is found with the given options.

Signed-off-by: Stacy Harper <contact@stacyharper.net>