make the hide-menu gesture double as an action for closing all notifications
I regularly had situations where I received a whole pile of
notifications that were hard to close. Dunst is configured to
open the notification upon click and close upon right click, but
we can't do a right click on our touchscreens, and I didn't want to
override the default left-click behaviour that might be useful.
This should hopefully solve it.

(this solution requires dbus-x11 to be installed)

Signed-off-by: Stacy Harper <contact@stacyharper.net>