Prevent the incomingcallmenu to go futher if the menu failed
This can happen if an call come while a menu already is open of if the
user kill the dmenu with a top side gesture.

We dont want the notification to be purged to allow the user to manually
open it with the notification appmenu so we have to stop the code here.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Anjandev Momi <anjan@momi.ca>