Remove -a cause it can break the system.
If the apk cache is outdated (first installation) and you try to upgrade
with this "-a" argument without configuring a network, then apk will
just strip every packages that are not available.

It then leave the phone in a very broken state.

The -a was used to handle mainlining software from pmaport to aport
(where pmaport uses 9999 version numbers). As we do not rely that much
on those softwares (that are now mainlined), we strip this unsafe
argument.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Anjandev Momi <anjan@momi.ca>