Allow the userscripts to be a file instead of a folder
At this moment, we list the files in .config/sxmo/userscripts/ to add
entries on the Userscripts submenu.

This way we cannot customize the entry label cause the script filename
will be the displayed entry. So we cannot use icons or emojies in the
label.

With this patch, we can use a plain text file located at
~/.config/sxmo/userscripts as appmenu entries, instead of a directory at
this same location.

This file can be formated like this:

 AddEvent ^ 0 ^ ~/bin/add_event
 Mail     ^ 0 ^ ~/bin/mail
 Weather  ^ 0 ^ sxmo_terminal.sh -f "monospace:size=5" sh -c "curl http://wttr.in/ | less -SR"

This allow the entry name to be different from the command it runs.
This also allow icons in the entries.
This also allow single line little scripts.

This patch still allow the current behavior with directory listing. The
location can be either a file, or a folder.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>