fix ssh mode since we dropped xdm-config
xdm-config was adding a /etc/profile.d sxmo script that was setting up
some env variabled. We now miss the XDG_DATA_HOME one that we was not
defaulting in sxmo_common.sh

As we prefer not to enforce those global variables, we will fix this
adding this default to the common script then. Scripts that need it
should load this script anyway.