Move modem/mms processing check from is_idle to can_suspend
In some case, we were preventing the phone to go to suspension properly
(if modem or mms was processing). It was very risky and random so we
have to find a better way to handle those cases.

The thing is that checking if an important process is running to not
suspend at this time is a good goal. But it should not be considered the
same as checking if the system is idle.

is_idle should check things that come from the user. It should not check
pseudo-random event like this.

This modem/mms check is more a "can I suspend now?" check and it should
be considered differently.

To check this without breaking is_idle behavior, we will just delay the
suspension on this case.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>