modem disconnected state is enabled
Maybe I'm misreading this documentation here:

https://www.freedesktop.org/software/ModemManager/doc/latest/ModemManager/ref-overview-modem-state-machine.html

I noticed the 'disconnected' icon never showed up. This apparently is
because its not defined...  When my carrier is out of range and modem
gives up (i.e., it is disconnected from carrier), it is set to
'enabled'.  Also added connecting/searching as options.  Could add
others?

Signed-off-by: Stacy Harper <contact@stacyharper.net>