Refactor modemsendsms and modemtext v2
This is not the final patch, but for those who are interested.  I
noticed two problems with our current MMS setup. (1) we should be
deleting mms as they come in, and (2) there was not a robust error
checking for failed sends.  This patch should fix that (it also adds
some debug stuff that'll be removed in the final version).

Comments on the approach are welcome and appreciated.

Signed-off-by: Stacy Harper <contact@stacyharper.net>