bluetooth/wifitoggle: fix to make things permanent
On most systems, the rfkill states are saved on reboot (see
/etc/init.d/rfkill in alpine, for instance).  However, if we
unload the modules, the the rfkill state would not be visible,
and so toggle would not be permanent.

One solution would be to add the modules to blacklist, but that seems
too complicated.  The simpler solution is to not unload/load the modules
at all.  After all, rfkill should block these modules from running
and/or sucking up battery, nad you really only need to unload the module
if it is buggy.  Maybe there's a reason we want to unload the module,
but I think that might be a user decision.

Signed-off-by: Stacy Harper <contact@stacyharper.net>