Use wob to display volume and brightness changes
Wob is a simple wayland bar that display itself when a new line value is
given.

It is simple and fast and looks better when playing with volume
and brightness.

This change does not impact the dwm environment.

I would like not to compute useable_width manually but the margin value
only separate the status bar to the wob bar. Right and left side are
touching the device sides without this explicit value.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>