core/sxmo_winit: clean up old sxmo session processes before starting
If sway crashes, or something else happens such that the stopsway
function and stop hook is not executed, there are a lot of leftover
processes from the previous session.

I renamed swaystop to swaycleanup in this patch, and call it when
starting sway and after sway has stopped. This should make sure that
there are no residual sxmo/sway processes from the previous session when
starting a new session.

Signed-off-by: Anjandev Momi <anjan@momi.ca>