Refact: leds usage
- Use one single sxmo_led.sh script

$ sxmo_led.sh blink red
$ sxmo_led.sh blink red blue

$ sxmo_led.sh set red 100
$ sxmo_led.sh set red 100 blue 0 green 100

This script got a simple mutex check to behave consistently while
spamed.

Blinking leds now return to initial state. Make it more easy to use
with long running leds. Plus it also solve the issue of invisible
blinking led if the color already is used.

- Free leds of sxmo lock state

The goal here is to not rely on long running leds for screen lock state.

The typical use case is when the user want to lock the screen while
playing a video. Having a blue led in the eyes really is annoying in
this kind of cases.

Another issue is when the phone is in crust, then the mixup of green
+ red give a not simple yellow. Or when we make a color to blink above
already used leds.

As sxmo now rely on an idle daemon that slowly move it to crust, having
long running leds isnt that mandatory now.

It will also make it more easy for other script to notice the user. By
example, we dont have that much merged leds now.

To avoid this, we would rely more on blinkind leds:

Move to unlock: blinking yellow
Move to lock: blinking blue
Move to off: blinking purple (blink periodically purple while in off)
Move to crust: blinking red

Additionnaly, I added a  icon in the status bar to make it more
explicit if the phone stay locked for a while

- Drop rtcwake that conflict with blinking led and are pretty useless

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>