sxmo_modemdial: refactor
modemdial doesn't have any long running child processes, so killing them
is totally unnecessary. Also fatelerror was never called after the phone
call was started so it the cleanup it did was unnecessary.

staceee edits:

This script was ugly and needed a rewrite. Aren did the first parts and
I completed this patch.