Remove go to lock when waking up from usb power
As wart_ said, we dont have an automatic flow from off to unlock when
the usb power is connected. So having this flow from crust to unlock
seems inconsistent.

As having this off -> unlock flow would be not trivial to do, and as
this crust -> unlock seems pretty useless, we remove it.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>