Fix unlocksim logic.
First, this removes the unlocksim hook and makes an sxmo_unlocksim.sh,
which you can also access from the menu.

Second, this cleans up the unlocksim logic.  Both go through the 'modem'
hook now as they should.

One consequence here is that our check to see if the modem is locked
uses mmcli -m any -K which returns a string rather than a number,
whereas the dbus-monitor sends a number to the modem hook.  Hence, I
made modem hook aware of either string or number to determine the state.

This also simplifies the boot process for the modem.

Signed-off-by: Stacy Harper <contact@stacyharper.net>