sxmo_migrate: new versioned migration
This patch ensures that user customisations to configuration files and
hooks can be more easily kept up to date when we make changes with new
sxmo releases. We need this because our hooks are modified both by us
and the user, to ensure the user has maximal flexibility and freedom in
adjusting the system to his/her needs; needing a mechanism like this
is the cost for that.

Each config file carries a 'configversion' comment that signals the version of
that particular configuration file. The version is unique for each file
and can be anything (as long it wasn't used before); I suggest a simple
incrementing integer.

Configurations that are out of date are moved out of the way when Sxmo
starts, presenting the system-default for that file instead. This
ensures that users always have a working system after an upgrade. Users
need to run sxmo_migrate.sh to 'merge' the changes.

sxmo_migrate.sh now has several modes (1st parameter):

* interactive: the default mode, shows diffs and prompts the user to
  merge changes for file that are out of date. Like before, but makes
  sure to use the file that were 'moved out of the way' earlier.
* all: same as interactive, but prompts for all files even if not
  out of date
* reset: removes all user hooks and reverts to default configurations, good
  option if a user is stuck somehow
* sync: this is a non-interactive mode automatically invoked from
  sxmo_[xw]init.sh, it takes care of making sure default
  configuration files exist and it checks the configversion, moving
  outdated configs out of the way in favour of defaults.
* state: returns the current state, i.e. lists the files
  that have been marked as needing revision.
  (does not sync by itself)

Signed-off-by: Stacy Harper <contact@stacyharper.net>