Display correct number in migrate notification
Using if here seemed to be consuming the exit status of sxmo_migrate so
it would display a notification that said 0 files were out of date.
Using || instead doesn't seem to have the same issue. I haven't found
docs on this, but I tested the method with ash, bash, dash, and ksh.

Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>