sxmo_migrate: more explicit choices in resolving migration differences
This patch offers four clear choices for migration and newly allows for
retaining the user-version but only updating the configversion (this
assumes the user verified), so the user doesn't need to do this in the
editor manually. There's also a clear "ignore" option in which nothing
happens at all and which is the default, this allows postponing
migration of certain files.

Moreover, when opening the editor, it also shows the diff for convenience.

It also fixes several bugs that were still present:
 * $EDITOR exit code wasn't respected
 * the former "No" option led to the file being re-migrated away on next
   start because configversion wasn't update.

Signed-off-by: Stacy Harper <contact@stacyharper.net>