Rework the modem call menu
- Simplify the menus code

We use new patterns to make it simple to read the code

- Use callaudiod

This script isnt pulseaudio/pipewire perfectly compatible. It doesnt
create nodes that are natives and managed by those audio daemons. But it
connect to thoses to know the audio setup and offer a simple cli to
manage basic commands as speaker toggle by example. It also handle ootb
the connected jack audio devices which make our life way simpler.

This solution isnt perfect but it is better than the megi script we use
atm that is pinephone specific.