Do not unset SWAYSOCK if not running on a sxmo device
If you run sway on a desktop, if you install sxmo-utils, then your
terminal emulators will not have SWAYSOCK available anymore.

To avoid this, we only unset SWAYSOCK if the sxmo.swaysock file is
present. We keep this behavior to ensure we unset SWAYSOCK when
toggeling from sway to dwm by example.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>