Rework sxmo states
Some devices need more flexible states. Wart is working on the
Pinenote and in this device, a locked and off mode looks useless and
wrong. We have to be able to offer more diversity with our states.

We was determining the current state based on some input device state.
Are they enabled/disabled ? This brings lot of constraints :

- We cannot have abstract states. That would be usefull to have
dedicated keybinds at some points. We could have very dedicated
workflows.
- We have to maintain complex code to checks tons of different input
devices. Pinenote brings a touchscreen but also a stylus. We have to
make crazy assomptions to determine we are in a "locked" state.

With recent patches from wart, lot of code already has been moved to
hooks. The hooks "unlock", "lock", "screenoff" got everything to enable
and disable input and output devices. The user then can adapt it to
their need.

We will keep our 3 different state with those hooks. We will also keep
some of the code remaining from sxmo_screenlock.sh to sxmo_suspend.sh.

Calling sxmo_suspend.sh place the device in suspension. This is not a
state, this is a time travel.

We now store the current state in a file SXMO_STATE. Checking the
current state is way faster and it really is perceptible while using the
phone.