Simple unlock/locker states for desktops environments
This is a simple toggle between unlocked (the default) and a locker
mode.

Locker mode is swaylock with automatic dpms after 10 seconds.

We use swaylockd for a more secure default. This also prevent us from
killing and stopping swaylock. For this simple destkop usage it looks
good. For future mobile environments we might need to think more on
this.

To use it, the user either have to expose SXMO_DEVICE_NAME=desktop in
its ~/.profile (depending on hostname by example) or to copy/symlink
those hooks in their default hook folder.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>