Make claiming or releasing mutexes atomic
If one sed process started before the previous one finished, the changes
of all but the last one to finish would be lost.

Also I think if the grep is run at the same time as another sed process
it will might generate an outdated line number, causing sed to remove
the wrong lock item.

This is a relatively crude method for avoiding these race conditions,
until a method that doesn't need to lock it's lock can be developed.

I also also added compatibility with gnu xargs

Signed-off-by: Stacy Harper <contact@stacyharper.net>