Add contextmenu for mepo to generate via mepo_ui_central_menu.sh
As of mepo 0.4, mepo's default provided script mepo_ui_central_menu.sh
can generate a list of scripts and hotkeys mepo is bundled with. This
matches the same menu mepo brings up for long press or by pressing the
menu button in the UI. We should use this logic so that the menu entries
match on sxmo's context menu as well.

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>