split mute_ring, discard, and missed_call
While the default settings are the same, having these point to each
other like that was (a) kind of weird and (b) kind of annoying if
someone wants to change *just* missed_call (to have it e-mail you, for
instance), they would have to change mute_ring and discard too.

I think we just split them up.

Signed-off-by: Stacy Harper <contact@stacyharper.net>