sxmo_status: don't attempt to find XDG_RUNTIME_DIR
jk_000 on irc had a issue with the led script hanging, during the
process of debugging it it seemed to be looking for the led lock in
~/.local/run.

I don't know if this is related, but it should be changed now that we
make sure to always initialize XDG_RUNTIME_DIR, and don't use
.local/run.

v1 of this patch caused the tests to fail because XDG_RUNTIME_DIR isn't
set in the test environment. This sets it to a temporary directory
before the test run.

Signed-off-by: Stacy Harper <contact@stacyharper.net>