Fix SXMO_ vars in profile.d
After updating, sxmo_migrate.sh in a ssh session was broken, with
a strange error:

    usr/share/sxmo/migrations/1.9.0.0.sh: cd: line 8: can't cd to /home/user/.config/.config/.config/sxmo/hooks/: No such file or directory

This patch fix the issue introduced by c3991d4.

I'm not sure why phartman changed this in a second second version of
the patch, it was correct in the first one. Maybe I'm missing something.

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>