statusbar: fix error with keyboard bat
Sometimes the keyboard battery will return "Not a tty" when you cat
/sys/class/power_suppply/ip5xxx-charger/capacity.  This caused PCT
to not be set, and so caused a load of errors in log and a blank in the
statusbar.  This patch is a stop-gap fix for that, reporting ERR.  I'm
not 100% sure what is at root the problem (presumably it is kernel
related).

Signed-off-by: Stacy Harper <contact@stacyharper.net>