avoid drift in clocks
This adds a sleep program that wakes up when the time (in seconds) is
divisible by a specific number, so for example passing 60 to it will
make it exit at the beginning of each minute.

This program is also able to exit when the system exits suspend (when
the kernel syncs the clock), which means we don't need to signal the
clocks anymore.

Signed-off-by: Stacy Harper <contact@stacyharper.net>