new SXMO_EG25 envvar specifies if eg25-manager needed
We might need to add more logic to control other modem managers in the
future but I think pinephone is the only one that requires eg25-manager
so lets do this for now.

This fixes modem support on non eg-25manager devices like the poco f1.
I no longer need to manually switch on the modemmonitor on startup and I
dont get the message that eg25-manager failed to start.

Signed-off-by: Stacy Harper <contact@stacyharper.net>