Pinephone-Pro profile cleanup
Fixed SXMO_TOUCHSCREEN_ID, added SXMO_MONITOR, and made a small change
to sxmo_hook_start so we don't have to make xorg confs for devices. This
ensures that orientation is correct on devices where DSI-1 may not be
the primary monitor by default. It may not be the cleanest fix, but at
minimum, it doesn't hurt anything if SXMO_MONITOR isn't set.

I removed OG pinephone specific sys_files. I am not sure what their
equivalents are on the pro, or if they are necessary, but since they
don't exist on the pro anyway, it doesn't hurt to remove them.

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>