appmenu: require exact match with picked
When selecting "YouTube" from the scripts menu "YouTube (audio)" was
being selected. This happens because the audio version is first in the
choices list so it was matched first when grepping for "YouTube".

This patch solves that by requiring the name of the entry to match
exactly, not just contain the selected item.

Signed-off-by: Anjandev Momi <anjan@momi.ca>