Add initial support for samsung-gt510
Hi all,

I just tested SWMO on my new samsung-gt510 [1]. SWMO is displayed but input via
buttons and touchscreen does not work.

With this patch the device is quite usable. Some functionality (like LEDs and
modem) is missing / not tested, but this could already be 80 % of the port.

Greetings and thank you for your great work
xengineering

[1] https://wiki.postmarketos.org/wiki/Samsung_Galaxy_Tab_A_9.7_2015_(samsung-gt510)

Signed-off-by: Anjandev Momi <anjan@momi.ca>