Clean up modem daemons
I think we should get rid of system daemon management completly, at
least from our core scripts. We still offer a hook that the user can
override on custom systems to restart modem daemons in case of crash.

We continue to manage vvmd and mmsd throught superd cause they are user
daemons and we choosed to use one daemon manager for that.