Add fallback hook for unknown contextmenu choice
This patch adds a new hook that's called when the user submits an
unknown choice to a contextmenu. This is useful to be able to
handle arbitrary text inputs beyond the listed choices. This can
be used to search files, do a web search, run commands, etc.

The hook takes two arguments: the $WINNAME of the contextmenu that
was used and the $CHOICE typed into the menu by the user. This
should allow fine-grained control of the fallback depending on
where the choice was entered.

In my personal config I currently use a simple script just to open
URLs or search the web. You can see it here:

https://git.sr.ht/~brecht/sxmo-config/tree/master/item/hooks/contextmenu_fallback
Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>