fix conflicting audio settings (#502)
I agree with the spirit, but I think the retry loops are needed. I added
some log output so we can track it for now.  I still wasn't sure about
the moving of unmute mic from beginning to end, so I wanted to send this
to the list before I apply it.

(Original Comment:
When callaudiod disables call audio mode it always enables the mic.
Because we were disabling the mic before disabling call audio mode
reset_audio always fails.

I also removed the retry loop; if the software we rely on breaks we
should fix it upstream, rather than hitting it with a hammer until it
works.

This should fix https://todo.sr.ht/~mil/sxmo-tickets/502
End Original Comment)

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>
Signed-off-by: Stacy Harper <contact@stacyharper.net>