Always prefer user hooks
This changes the order of hook dirs in the path so that device agnostic
user hooks will be preferred over device specific default hooks.

I expect most users aren't sharing hooks between different types of
device, this should make it easier for them to just copy a hook.

Users with device specific hooks will still need to be careful, but
that's already the case without this patch.

Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>